Add map importing + exporting support for the .ffs format. The .ffs format is new and will have a blender plugin soon.

+++ src/net/highwayfrogs/editor/file/map/animation/MAPAnimation.java
@@ -234,7 +234,7 @@ public class MAPAnimation extends GameObject {
         }
 
         editor.addButton("Add Texture", () -> {
-            getTextures().add(getTextures().isEmpty() ? 0 : getTextures().get(getTextures().size() - 1));
+            getTextures().add(getTextures().isEmpty() ? (short) 0 : getTextures().get(getTextures().size() - 1));
             manager.setupEditor();
         });
     }
